requirements.txt
numpy>=1.24,<3
pandas>=2.0
pyarrow>=12
scikit-learn>=1.3,<2
statsmodels>=0.14
lightgbm>=4.0
matplotlib>=3.7
requests>=2.31
ruamel.yaml>=0.17
